Rename Endpoint to EndpointLegacy and create new Endpoint annotation
