+++ sample/src/main/java/es/sandwatch/trim/sample/Sample.java
@@ -1,7 +1,7 @@
 package es.sandwatch.trim.sample;
 
 import es.sandwatch.trim.ApiSpecification;
-import es.sandwatch.trim.Endpoint;
+import es.sandwatch.trim.EndpointLegacy;
 import es.sandwatch.trim.Report;
 import es.sandwatch.trim.Trim;
 import es.sandwatch.trim.sample.model.Category;
@@ -18,14 +18,14 @@ import org.jetbrains.annotations.NotNull;
 public class Sample implements Trim.ProgressListener{
     private Sample(){
         ApiSpecification spec = new ApiSpecification()
-                .addEndpoint(new Endpoint("http://app.tndata.org/api/categories/23/", Category.class))
-                .addEndpoint(new Endpoint("http://app.tndata.org/api/goals/82/", Goal.class));
+                .addEndpoint(new EndpointLegacy("http://app.tndata.org/api/categories/23/", Category.class))
+                .addEndpoint(new EndpointLegacy("http://app.tndata.org/api/goals/82/", Goal.class));
         Report report = Trim.run(spec, this);
         System.out.println(report);
     }
 
     @Override
-    public void onEndpointReportComplete(@NotNull Endpoint endpoint, int completed){
+    public void onEndpointReportComplete(@NotNull EndpointLegacy endpoint, int completed){
         System.out.println(endpoint + ", " + completed + " completed so far.");
     }
